fix: homepage 경로 변경
